Move allocate/storage to storage/allocate
