+++ packages/codec/lib/evm/types.ts
@@ -1,7 +1,7 @@
 import * as Common from "@truffle/codec/common";
 import * as Storage from "@truffle/codec/storage/types";
 import * as Ast from "@truffle/codec/ast";
-import { StorageAllocations } from "@truffle/codec/allocate/storage/types";
+import { StorageAllocations } from "@truffle/codec/storage/allocate/types";
 import { MemoryAllocations } from "@truffle/codec/memory/allocate/types";
 import {
   AbiAllocations,
