+++ packages/codec/lib/storage/index.ts
@@ -3,5 +3,5 @@ export * from "./types";
 import * as Utils from "./utils";
 export { Utils };
 
-import * as Allocate from "@truffle/codec/allocate/storage";
+import * as Allocate from "./allocate";
 export { Allocate };
